requirements.txt
# Use CPU-only PyTorch to reduce image size (no GPU on Railway)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.0+cpu

boto3==1.34.34
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
pinecone-client>=3.0.0
langchain-pinecone>=0.1.0
sentence-transformers==2.3.1
pdfplumber==0.10.3
PyPDF2==3.0.1
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
python-dotenv==1.0.0
tqdm==4.66.1
langchain-groq
pytesseract
pillow
spacy
sentence-transformers>=2.3.0
langchain-huggingface
langchain-text-splitters
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
slowapi>=0.1.9
langchain_pinecone>=0.1.0
pinecone-client